refactor: assets - moved the `animated-night.svg` in the assets folder for the dashboard - updated the paths in the login and register forms
